valorDolar/requirements.txt
django==2.1
djangorestframework==3.8.2
beautifulsoup4==4.6.1
requests==2.19.1
